Move CheckState
