+++ java/system-test/system-test-suite/src/test/java/sleeper/systemtest/suite/investigate/CheckState.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.systemtest.suite;
+package sleeper.systemtest.suite.investigate;
 
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDB;
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDBClientBuilder;
